fix: report archive sizes in bytes, and stop hiding small packs
Turns storeReportsSizeInBytesAsItsNameAndDocsSay green. app 211 tests / 0
failures, frontend 35 / 0 (was 32 -- the three new ones are the formatter).

size was computed as size().div(1048576.0).toInt() while SavedFile, ModPack and
ServerPack all documented it as bytes. Everything under a mebibyte therefore
reported 0, and both pack tables gated their download button on
`v-if="props.row.size > 0"` -- so a modpack or server pack under 1 MiB was not
merely mislabelled, it could not be downloaded from the tables at all. Server
packs are routinely that small.

The field widens to Long with the unit, necessarily: the shipped
spring.servlet.multipart.max-file-size is 5000MB, which does not fit in an Int,
so storing bytes in the old type would overflow.

Formatting moves to the SPA, where the unit is known: src/utils/format.ts renders
binary units and answers "unknown" for an absent size rather than
"NaN undefined". Both tables and both cards use it, and both `size > 0` gates are
gone rather than re-expressed -- the tables stay untested, per the standing
decision recorded in the frontend's CLAUDE.md, so the logic that could be tested
was moved out of them. Its guards were verified by mutation: scaling at 1048576
instead of 1024 fails two cases, dropping the non-finite guard fails the third.

Existing documents are deliberately not rewritten. A value-only migration cannot
tell a converted row from an unconverted one -- nothing marks them, the same trap
already recorded for supersededLegacyKey -- and it no longer matters: with the
size > 0 gate gone the field is cosmetic, so an old row simply renders as a few
hundred bytes. Stated in claude-docs/API-BEHAVIOUR-CHANGES.md rather than
silently accepted.

Response samples in Modpacks.md and Server-Packs.md restated in bytes.
api-docs.yaml still says int32 and is regenerated in the next commit, together
with the multipart request body it also gets wrong.

+++ serverpackcreator-app/src/main/kotlin/de/griefed/serverpackcreator/app/web/serverpack/ServerPack.kt
@@ -42,7 +42,7 @@ class ServerPack {
     /** Which [ModPack][de.griefed.serverpackcreator.app.web.modpack.ModPack] this was generated from. */
     var modpackId: String = ""
     /** Archive size in bytes. */
-    var size: Int = 0
+    var size: Long = 0
     /** How often this server pack has been downloaded. */
     var downloads: Int = 0
     /** How many users have voted that this pack actually runs — a counter, not a flag. */
@@ -61,7 +61,7 @@ class ServerPack {
     var runConfiguration: RunConfiguration? = null
 
     constructor(
-        size: Int,
+        size: Long,
         runConfiguration: RunConfiguration?,
         fileID: String?,
         fileName: String?,
@@ -79,7 +79,7 @@ class ServerPack {
     @PersistenceCreator
     private constructor(
         id: String,
-        size: Int,
+        size: Long,
         downloads: Int,
         confirmedWorking: Int,
         dateCreated: Date,
